just open browser

+++ fiberwebgui.go
@@ -78,29 +78,33 @@ func getFreePort() int {
 	return port
 }
 
-func startBrowser(guiWg *sync.WaitGroup, browserClosed chan bool, browserPath, port string, windowSizeFlag string) {
-	tempDir, _ := os.MkdirTemp("", "fiberwebgui")
-
+func startBrowser(guiWg *sync.WaitGroup, browserClosed chan bool, browserPath, port string, windowSizeFlag string, appMode bool) {
 	url := "http://127.0.0.1:" + port
 	browserExecPath := browserPath
-	userDataDir := "--user-data-dir=" + tempDir
-	newWindow := "--new-window"
-	noFirstRun := "--no-first-run"
-	appUrl := "--app=" + url
 
-	log.Println("Browser started with: ", browserExecPath, userDataDir, newWindow, noFirstRun, windowSizeFlag, appUrl)
-
-	cmd := exec.Command(browserExecPath, userDataDir, newWindow, noFirstRun, windowSizeFlag, appUrl)
-	err := cmd.Run()
-	if err != nil {
-		log.Fatal(err)
+	if appMode {
+		tempDir, _ := os.MkdirTemp("", "fiberwebgui")
+		userDataDir := "--user-data-dir=" + tempDir
+		newWindow := "--new-window"
+		noFirstRun := "--no-first-run"
+		appUrl := "--app=" + url
+		log.Println("Browser started with: ", browserExecPath, userDataDir, newWindow, noFirstRun, windowSizeFlag, appUrl)
+		cmd := exec.Command(browserExecPath, userDataDir, newWindow, noFirstRun, windowSizeFlag, appUrl)
+		err := cmd.Run()
+		if err != nil {
+			log.Fatal(err)
+		}
+		os.RemoveAll(tempDir)
+		log.Println("Browser stopped!")
+		browserClosed <- true
+		guiWg.Done()
+	} else {
+		cmd := exec.Command(browserExecPath, url)
+		err := cmd.Run()
+		if err != nil {
+			log.Fatal(err)
+		}
 	}
-
-	os.RemoveAll(tempDir)
-	log.Println("Browser stopped!")
-	browserClosed <- true
-	guiWg.Done()
-
 }
 
 func startFiberServer(guiWg *sync.WaitGroup, browserClosed chan bool, app FiberApp, port string) {
@@ -118,33 +122,44 @@ func startFiberServer(guiWg *sync.WaitGroup, browserClosed chan bool, app FiberA
 
 }
 
-func run(app FiberApp, windowSizeFlag string, port int) {
+func run(app FiberApp, windowSizeFlag string, port int, appMode bool) {
 	browserPath := getBrowserPath()
 	portstr := strconv.Itoa(port)
+
 	browserClosed := make(chan bool)
 	var guiWg sync.WaitGroup
 	guiWg.Add(2)
-	go startBrowser(&guiWg, browserClosed, browserPath, portstr, windowSizeFlag)
+	go startBrowser(&guiWg, browserClosed, browserPath, portstr, windowSizeFlag, appMode)
 	go startFiberServer(&guiWg, browserClosed, app, portstr)
 	guiWg.Wait()
+
 }
 
 func Run(app FiberApp) {
 	port := getFreePort()
-	run(app, "--start-maximized", port)
+	run(app, "--start-maximized", port, true)
 }
 
 func RunOnPort(app FiberApp, port int) {
-	run(app, "--start-maximized", port)
+	run(app, "--start-maximized", port, true)
 }
 
 func RunWithSize(app FiberApp, width, height int) {
 	port := getFreePort()
 	windowSize := "--window-size=" + strconv.Itoa(width) + "," + strconv.Itoa(height)
-	run(app, windowSize, port)
+	run(app, windowSize, port, true)
 }
 
 func RunWithSizeOnPort(app FiberApp, width, height, port int) {
 	windowSize := "--window-size=" + strconv.Itoa(width) + "," + strconv.Itoa(height)
-	run(app, windowSize, port)
+	run(app, windowSize, port, true)
+}
+
+func RunBrowser(app FiberApp) {
+	port := getFreePort()
+	run(app, "", port, false)
+}
+
+func RunBrowserOnPort(app FiberApp, port int) {
+	run(app, "", port, false)
 }
